refactor: :recycle: separate fiat selection, infobox from header
