+++ src/components/index.ts
@@ -2,4 +2,6 @@ export { default as Header } from './Header'
 export { default as Layout } from './Layout'
 export { default as FeeTable } from './FeeTable'
 export { default as Chart } from './Chart'
+export { default as Infobox } from './Infobox'
+export { default as FiatSelection } from './FiatSelection'
 export { default as TxnSizeSlider } from './TxnSizeSlider'
